+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `auto_increment` to `receiptComponents` type
+- `startingAutoIncrement` to `externalIdGeneration` configuration
+
 ## [9.1.0] - 2024-10-09
 
 ### Added